Get rid of mypy errors.
We need to ignore numpy and to heavily use assertions/exceptions to down-scope
the types we actually use without violating the Liskov principle.
This wouldn't be necessary if we had proper metaclasses :/ maybe we should look
into ABC.

The List[B] -> Sequence[B] changes are because Sequences, unlike Lists are
covariant and allow specializing the type of their elements.
This is WAI https://github.com/python/mypy/issues/2984